+++ app-frontend/src/app/components/pages/project/layers/index.js
@@ -1,11 +1,20 @@
 import _ from 'lodash';
 import tpl from './index.html';
-import {Set, Map} from 'immutable';
+import { Set, Map } from 'immutable';
 
 class ProjectLayersPageController {
     constructor(
-        $rootScope, $state, $q,
-        projectService, paginationService, modalService, authService, mapService
+        $rootScope,
+        $state,
+        $q,
+        $log,
+        $scope,
+        projectService,
+        paginationService,
+        modalService,
+        authService,
+        mapService,
+        projectEditService
     ) {
         'ngInject';
         $rootScope.autoInject(this, arguments);
@@ -13,22 +22,21 @@ class ProjectLayersPageController {
 
     $onInit() {
         this.selected = new Map();
-        this.visible = new Set([this.project.defaultLayerId]);
+        const visibleLayers = this.projectEditService.getVisibleProjectLayers();
+        this.visible = visibleLayers.size ? visibleLayers : Set([this.project.defaultLayerId]);
         this.syncMapLayersToVisible();
-        this.projectService.getProjectPermissions(this.project, this.user).then(
-            permissions => {
-                this.permissions = permissions.map(p => p.actionType);
-            });
-        this.projectService.getProjectLayerStats(this.project.id).then(
-            layerSceneCounts => {
-                this.layerStats = layerSceneCounts;
-            });
+        this.projectService.getProjectPermissions(this.project, this.user).then(permissions => {
+            this.permissions = permissions.map(p => p.actionType);
+        });
+        this.projectService.getProjectLayerStats(this.project.id).then(layerSceneCounts => {
+            this.layerStats = layerSceneCounts;
+        });
         this.fetchPage();
     }
 
     $onDestroy() {
         // remove layers from map
-        this.getMap().then((map) => {
+        this.getMap().then(map => {
             map.deleteLayers('Project Layers');
         });
     }
@@ -44,41 +52,46 @@ class ProjectLayersPageController {
     fetchPage(page = this.$state.params.page || 1) {
         this.layerList = [];
         this.layerActions = {};
-        const currentQuery = this.projectService.getProjectLayers(
-            this.project.id,
-            {
+        const currentQuery = this.projectService
+            .getProjectLayers(this.project.id, {
                 pageSize: 30,
                 page: page - 1
-            }
-        ).then((paginatedResponse) => {
-            this.layerList = paginatedResponse.results;
-            this.layerList.forEach((layer) => {
-                layer.subtext = '';
-                if (layer.id === this.project.defaultLayerId) {
-                    layer.subtext += 'Default layer';
+            })
+            .then(
+                paginatedResponse => {
+                    this.layerList = paginatedResponse.results;
+                    this.layerList.forEach(layer => {
+                        layer.subtext = '';
+                        if (layer.id === this.project.defaultLayerId) {
+                            layer.subtext += 'Default layer';
+                        }
+                        if (layer.smartLayerId) {
+                            layer.subtext += layer.subtext.length ? ', Smart layer' : 'Smart Layer';
+                        }
+                    });
+                    const defaultLayer = this.layerList.find(
+                        l => l.id === this.project.defaultLayerId
+                    );
+                    this.layerActions = this.layerList.map(l =>
+                        this.addLayerActions(l, defaultLayer === l)
+                    );
+                    this.pagination = this.paginationService.buildPagination(paginatedResponse);
+                    this.paginationService.updatePageParam(page);
+                    if (this.currentQuery === currentQuery) {
+                        delete this.fetchError;
+                    }
+                },
+                e => {
+                    if (this.currentQuery === currentQuery) {
+                        this.fetchError = e;
+                    }
                 }
-                if (layer.smartLayerId) {
-                    layer.subtext += layer.subtext.length ? ', Smart layer' : 'Smart Layer';
+            )
+            .finally(() => {
+                if (this.currentQuery === currentQuery) {
+                    delete this.currentQuery;
                 }
             });
-            const defaultLayer = this.layerList.find(l => l.id === this.project.defaultLayerId);
-            this.layerActions = this.layerList.map(
-                (l) => this.addLayerActions(l, defaultLayer === l)
-            );
-            this.pagination = this.paginationService.buildPagination(paginatedResponse);
-            this.paginationService.updatePageParam(page);
-            if (this.currentQuery === currentQuery) {
-                delete this.fetchError;
-            }
-        }, (e) => {
-            if (this.currentQuery === currentQuery) {
-                this.fetchError = e;
-            }
-        }).finally(() => {
-            if (this.currentQuery === currentQuery) {
-                delete this.currentQuery;
-            }
-        });
         this.currentQuery = currentQuery;
         return currentQuery;
     }
@@ -104,7 +117,8 @@ class ProjectLayersPageController {
                 {
                     icon: 'icon-eye',
                     isActive: () => this.visible.has(layer.id)
-                }, {
+                },
+                {
                     icon: 'icon-eye-off',
                     isActive: () => !this.visible.has(layer.id)
                 }
@@ -116,10 +130,8 @@ class ProjectLayersPageController {
         };
         const editAction = {
             name: 'Edit',
-            callback: () => this.$state.go(
-                'project.layer',
-                {projectId: this.project.id, layerId: layer.id}
-            ),
+            callback: () =>
+                this.$state.go('project.layer', { projectId: this.project.id, layerId: layer.id }),
             menu: true
         };
         const setDefaultAction = {
@@ -134,44 +146,49 @@ class ProjectLayersPageController {
         };
         const publishAction = {
             name: 'Publishing',
-            callback: () => this.$state.go(
-                'project.layer.settings.publishing',
-                {projectId: this.project.id, layerId: layer.id}
-            ),
+            callback: () =>
+                this.$state.go('project.layer.settings.publishing', {
+                    projectId: this.project.id,
+                    layerId: layer.id
+                }),
             menu: true
         };
         const exportAction = {
             name: 'Export',
-            callback: () => this.$state.go(
-                'project.layer.settings.publishing',
-                {projectId: this.project.id, layerId: layer.id}
-            ),
+            callback: () =>
+                this.$state.go('project.layer.settings.publishing', {
+                    projectId: this.project.id,
+                    layerId: layer.id
+                }),
             menu: true
         };
         const importAction = {
             name: 'Import imagery',
-            callback: () => this.modalService.open({
-                component: 'rfSceneImportModal',
-                resolve: {
-                    project: () => this.project,
-                    layer: () => layer,
-                    origin: () => 'project'
-                }
-            })
+            callback: () =>
+                this.modalService.open({
+                    component: 'rfSceneImportModal',
+                    resolve: {
+                        project: () => this.project,
+                        layer: () => layer,
+                        origin: () => 'project'
+                    }
+                })
         };
         const browseAction = {
             name: 'Browse for imagery',
-            callback: () => this.$state.go(
-                'project.layer.browse',
-                {projectId: this.project.id, layerId: layer.id}
-            )
+            callback: () =>
+                this.$state.go('project.layer.browse', {
+                    projectId: this.project.id,
+                    layerId: layer.id
+                })
         };
         const settingsAction = {
             name: 'Settings',
-            callback: () => this.$state.go(
-                'project.layer.settings',
-                {projectId: this.project.id, layerId: layer.id}
-            ),
+            callback: () =>
+                this.$state.go('project.layer.settings', {
+                    projectId: this.project.id,
+                    layerId: layer.id
+                }),
             menu: true
         };
         const deleteAction = {
@@ -186,19 +203,23 @@ class ProjectLayersPageController {
         // TODO: add implement these when the views are finished:
         // const unimplementedActions = [publishAction, exportAction, settingsAction];
         const layerActions = [
-            previewAction, editAction, editAoiAction, createAnalysisAction,
-            importAction, browseAction
+            previewAction,
+            editAction,
+            editAoiAction,
+            createAnalysisAction,
+            importAction,
+            browseAction
         ];
 
         return [
-            ...!_.get(layer, 'geometry.type') ? [alertAoiAction] : [],
+            ...(!_.get(layer, 'geometry.type') ? [alertAoiAction] : []),
             ...layerActions,
-            ...!isDefaultLayer ? [setDefaultAction, deleteAction] : []
+            ...(!isDefaultLayer ? [setDefaultAction, deleteAction] : [])
         ];
     }
 
     allVisibleSelected() {
-        let layerSet = new Set(this.layerList.map(l => l.id));
+        let layerSet = Set(this.layerList.map(l => l.id));
         return layerSet.intersect(this.selected.keySeq()).size === layerSet.size;
     }
 
@@ -206,9 +227,7 @@ class ProjectLayersPageController {
         if (this.allVisibleSelected()) {
             this.selected = this.selected.clear();
         } else {
-            this.selected = this.selected.merge(
-                new Map(this.layerList.map(i => [i.id, i]))
-            );
+            this.selected = this.selected.merge(new Map(this.layerList.map(i => [i.id, i])));
         }
         this.updateSelectText();
     }
@@ -240,15 +259,20 @@ class ProjectLayersPageController {
         } else {
             this.visible = this.visible.add(id);
         }
+        this.projectEditService.setVisibleProjectLayers(this.visible);
         this.syncMapLayersToVisible();
     }
 
     setProjectDefaultLayer(layer) {
-        this.projectService.updateProject(Object.assign({}, this.project, {
-            defaultLayerId: layer.id
-        })).then(() => {
-            this.$state.go('.', {}, {inherit: true, reload: true});
-        });
+        this.projectService
+            .updateProject(
+                Object.assign({}, this.project, {
+                    defaultLayerId: layer.id
+                })
+            )
+            .then(() => {
+                this.$state.go('.', {}, { inherit: true, reload: true });
+            });
     }
 
     createAnalysis(layer) {
@@ -274,9 +298,9 @@ class ProjectLayersPageController {
                     title: () => `Really delete ${ids.length} layers?`,
                     subtitle: () => 'Deleting layers cannot be undone',
                     content: () =>
-                        '<h2>Do you wish to continue?</h2>'
-                        + '<p>Future attempts to access these '
-                        + 'layers or associated annotations, tiles, and scenes will fail.',
+                        '<h2>Do you wish to continue?</h2>' +
+                        '<p>Future attempts to access these ' +
+                        'layers or associated annotations, tiles, and scenes will fail.',
                     feedbackIconType: () => 'danger',
                     feedbackIcon: () => 'icon-warning',
                     feedbackBtnType: () => 'btn-danger',
@@ -291,9 +315,9 @@ class ProjectLayersPageController {
                     title: () => 'Really delete layer?',
                     subtitle: () => 'Deleting layers cannot be undone',
                     content: () =>
-                        '<h2>Do you wish to continue?</h2>'
-                        + '<p>Future attempts to access this '
-                        + 'layer or associated annotations, tiles, and scenes will fail.',
+                        '<h2>Do you wish to continue?</h2>' +
+                        '<p>Future attempts to access this ' +
+                        'layer or associated annotations, tiles, and scenes will fail.',
                     feedbackIconType: () => 'danger',
                     feedbackIcon: () => 'icon-warning',
                     feedbackBtnType: () => 'btn-danger',
@@ -302,30 +326,39 @@ class ProjectLayersPageController {
                 }
             }).result;
         }
-        modal.then(() => {
-            const promises = ids.map(
-                id => this.projectService.deleteProjectLayer(this.project.id, id)
-            );
-            this.$q.all(promises).then(() => {
-                this.visible = this.visible.subtract(this.selected.keySeq());
-                this.selected = new Map();
-            }).catch(e => {
-                this.$log.error(e);
-            }).finally(() => {
-                this.fetchPage();
+        modal
+            .then(() => {
+                const promises = ids.map(id =>
+                    this.projectService.deleteProjectLayer(this.project.id, id)
+                );
+                this.$q
+                    .all(promises)
+                    .then(() => {
+                        this.visible = this.visible.subtract(this.selected.keySeq());
+                        this.projectEditService.setVisibleProjectLayers(this.visible);
+                        this.selected = new Map();
+                    })
+                    .catch(e => {
+                        this.$log.error(e);
+                    })
+                    .finally(() => {
+                        this.fetchPage();
+                    });
+            })
+            .catch(() => {
+                // modal closed
             });
-        }).catch(() => {
-            // modal closed
-        });
     }
 
     showDefaultLayer() {
-        this.visible = new Set([this.project.defaultLayerId]);
+        this.visible = Set([this.project.defaultLayerId]);
+        this.projectEditService.setVisibleProjectLayers(this.visible);
         this.syncMapLayersToVisible();
     }
 
     showPageLayers() {
         this.visible = this.visible.union(this.layerList.map(l => l.id));
+        this.projectEditService.setVisibleProjectLayers(this.visible);
         this.syncMapLayersToVisible();
     }
 
@@ -355,9 +388,7 @@ class ProjectLayersPageController {
             }
         });
 
-        modal.result
-            .then(() => this.fetchPage())
-            .catch(() => {});
+        modal.result.then(() => this.fetchPage()).catch(() => {});
     }
 
     createItemInfo(layer) {
@@ -396,7 +427,7 @@ class ProjectLayersPageController {
     }
 
     goToAoiDef(id) {
-        this.$state.go('project.layer.aoi', {layerId: id, projectId: this.project.id});
+        this.$state.go('project.layer.aoi', { layerId: id, projectId: this.project.id });
     }
 }
 
@@ -413,5 +444,4 @@ const component = {
 export default angular
     .module('components.pages.projects.layers', [])
     .controller(ProjectLayersPageController.name, ProjectLayersPageController)
-    .component('rfProjectLayersPage', component)
-    .name;
+    .component('rfProjectLayersPage', component).name;